refacto: création d'un hook teams.ts pour gérer les actions CRUD des teams et séparer la logique du bouton
